@@ -10,29 +10,27 @@ class JpegOptim extends AbstractCompressor
   /**
    * @var string|null
    */
-  private $_command;
+  private mixed $_command;
 
 
   /**
    * JpegOptim constructor.
    */
   public function __construct() {
-    $this->_command = array_key_exists('CMD_JPEGOPTIM', $_ENV)
-      ? $_ENV['CMD_JPEGOPTIM']
-      : null;
+    $this->_command = $_ENV['CMD_JPEGOPTIM'] ?? null;
   }
 
   /**
    * @inheritDoc
    */
-  function canCompress(string $format) {
+  function canCompress(string $format): bool {
     return !is_null($this->_command) && in_array($format, ['jpg', 'jpeg']);
   }
 
   /**
    * @inheritDoc
    */
-  function compress(string $fileName) {
+  function compress(string $fileName): bool {
     shell_exec(escapeshellcmd(
       implode(' ', array_map('escapeshellarg', [
         $this->_command,
