@@ -10,30 +10,28 @@ class OptiPng extends AbstractCompressor
   /**
    * @var string|null
    */
-  private $_command;
+  private mixed $_command;
 
 
   /**
    * OptiPng constructor.
    */
   public function __construct() {
-    $this->_command = array_key_exists('CMD_OPTIPNG', $_ENV)
-      ? $_ENV['CMD_OPTIPNG']
-      : null;
+    $this->_command = $_ENV['CMD_OPTIPNG'] ?? null;
   }
 
   /**
    * @inheritDoc
    */
-  function canCompress(string $format) {
-    return !is_null($this->_command) && in_array($format, ['png']);
+  function canCompress(string $format): bool {
+    return !is_null($this->_command) && $format == 'png';
   }
 
   /**
    * @inheritDoc
    */
-  function compress(string $fileName) {
-    $result = shell_exec($cmd = escapeshellcmd(
+  function compress(string $fileName): bool {
+    shell_exec(escapeshellcmd(
       implode(' ', array_map('escapeshellarg', [
         $this->_command,
         $fileName,
