@@ -10,39 +10,39 @@ class LanguageGroup
   /**
    * @var string[]
    */
-  private $languages = array();
+  private array $_languages = [];
 
   /**
    * @var float
    */
-  private $quality;
+  private float $_quality;
 
 
   /**
    * LanguageGroup constructor.
    * @param float $quality
    */
-  public function __construct($quality) {
-    $this->quality = $quality;
+  public function __construct(float $quality) {
+    $this->_quality = $quality;
   }
 
   /**
-   * @param $language
+   * @param string $language
    */
-  public function addLanguage($language) {
-    $this->languages[] = $language;
+  public function addLanguage(string $language): void {
+    $this->_languages[] = $language;
   }
 
   /**
    * @param LanguageStack $result
    * @param LanguageGroup $otherGroup
    */
-  public function combine(LanguageStack $result, LanguageGroup $otherGroup) {
-    foreach ($otherGroup->languages as $otherLanguage) {
+  public function combine(LanguageStack $result, LanguageGroup $otherGroup): void {
+    foreach ($otherGroup->_languages as $otherLanguage) {
       $isWildcard = $otherLanguage === '*';
       $quality = $otherGroup->getQuality() * $this->getQuality();
 
-      foreach ($this->languages as $language) {
+      foreach ($this->_languages as $language) {
         $score = $isWildcard
           ? 1
           : self::languageScore($otherLanguage, $language);
@@ -57,32 +57,36 @@ class LanguageGroup
   /**
    * @return string
    */
-  public function getBestLanguage() {
-    return reset($this->languages);
+  public function getBestLanguage(): string {
+    return reset($this->_languages);
   }
 
   /**
    * @return float
    */
-  public function getQuality() {
-    return $this->quality;
+  public function getQuality(): float {
+    return $this->_quality;
   }
+
+
+  // Private methods
+  // ---------------
 
   /**
    * Compare two language tags and distinguish the degree of matching.
    *
-   * @param string $a
-   * @param string $b
+   * @param string $lft
+   * @param string $rgt
    * @return float
    */
-  private static function languageScore($a, $b) {
-    $a = explode('-', $a);
-    $b = explode('-', $b);
+  private static function languageScore(string $lft, string $rgt): float {
+    $lft = explode('-', $lft);
+    $rgt = explode('-', $rgt);
 
-    for ($index = 0, $n = min(count($a), count($b)); $index < $n; $index++) {
-      if ($a[$index] !== $b[$index]) break;
+    for ($index = 0, $n = min(count($lft), count($rgt)); $index < $n; $index++) {
+      if ($lft[$index] !== $rgt[$index]) break;
     }
 
-    return $index === 0 ? 0 : (float)$index / count($a);
+    return $index === 0 ? 0 : (float)$index / count($lft);
   }
 }
