@@ -22,32 +22,32 @@ abstract class AbstractQuery extends BaseObject
   /**
    * @var int
    */
-  public $pageSize = 20;
+  public int $pageSize = 20;
 
   /**
    * @var string
    */
-  public $sort;
+  public string $sort;
 
   /**
-   * @var boolean
+   * @var string
    */
-  public $sortDirection = 'asc';
+  public string $sortDirection = 'asc';
 
   /**
    * @var AbstractFilter[]
    */
-  protected $_filters;
+  protected array $_filters;
 
   /**
    * @var Paginator
    */
-  protected $_paginator;
+  protected Paginator $_paginator;
 
   /**
    * @var ElementQuery
    */
-  protected $_query;
+  protected ElementQuery $_query;
 
   /**
    * @var string[]
@@ -83,17 +83,17 @@ abstract class AbstractQuery extends BaseObject
    * @param string $sort
    * @return bool
    */
-  public function allowSort(string $sort) {
+  public function allowSort(string $sort): bool {
     return in_array($sort, static::ALLOWED_SORTS);
   }
 
   /**
    * @return string
    */
-  public function getBasePath() {
+  public function getBasePath(): string {
     try {
       return Craft::$app->getRequest()->getPathInfo();
-    } catch (Throwable $error) {
+    } catch (Throwable) {
       return '';
     }
   }
@@ -101,23 +101,25 @@ abstract class AbstractQuery extends BaseObject
   /**
    * @param array $config
    * @return ToolbarDisplay
+   * @noinspection PhpUnused
    */
-  public function getToolbar(array $config = []) {
+  public function getToolbar(array $config = []): ToolbarDisplay {
     return new ToolbarDisplay($this, $config);
   }
 
   /**
    * @return AbstractFilter[]
    */
-  public function getFilters() {
+  public function getFilters(): array {
     return $this->_filters;
   }
 
   /**
    * @param string $key
    * @return string
+   * @noinspection PhpUnused
    */
-  public function getSortUrl(string $key) {
+  public function getSortUrl(string $key): string {
     if ($this->sort == $key) {
       return $this->getUrl([], [
         'sort' => $key,
@@ -135,15 +137,16 @@ abstract class AbstractQuery extends BaseObject
    * @param array $config
    * @return PaginationDisplay
    * @throws Throwable
+   * @noinspection PhpUnused
    */
-  public function getPagination($config = []) {
+  public function getPagination(array $config = []): PaginationDisplay {
     return new PaginationDisplay($this, $config);
   }
 
   /**
    * @return Paginator
    */
-  public function getPaginator() {
+  public function getPaginator(): Paginator {
     if (!isset($this->_paginator)) {
       $this->_paginator = new Paginator($this->_query, [
         'currentPage' => Craft::$app->getRequest()->getPageNum(),
@@ -159,7 +162,7 @@ abstract class AbstractQuery extends BaseObject
    * @param array $params
    * @return array
    */
-  public function getParameters($overrides = [], $params = []) {
+  public function getParameters(array $overrides = [], array $params = []): array {
     foreach ($this->_filters as $filter) {
       $params = array_merge($params, $filter->getParameters());
     }
@@ -169,15 +172,17 @@ abstract class AbstractQuery extends BaseObject
 
   /**
    * @return array
+   * @noinspection PhpUnused
    */
-  public function getResults() {
+  public function getResults(): array {
     return $this->getPaginator()->getPageResults();
   }
 
   /**
-   * @return float|int
+   * @return int
+   * @noinspection PhpUnused
    */
-  public function getTotalResults() {
+  public function getTotalResults(): int {
     return intval($this->getPaginator()->getTotalResults());
   }
 
@@ -186,7 +191,7 @@ abstract class AbstractQuery extends BaseObject
    * @param array $params
    * @return string
    */
-  public function getUrl($overrides = [], $params = []) {
+  public function getUrl(array $overrides = [], array $params = []): string {
     return Url::compose(
       $this->getBasePath(),
       $this->getParameters($overrides, $params)
@@ -196,14 +201,15 @@ abstract class AbstractQuery extends BaseObject
   /**
    * @return bool
    */
-  public function hasPagination() {
+  public function hasPagination(): bool {
     return $this->getPaginator()->getTotalPages() > 1;
   }
 
   /**
    * @return bool
+   * @noinspection PhpUnused
    */
-  public function hasResults() {
+  public function hasResults(): bool {
     return $this->getPaginator()->totalResults > 0;
   }
 
@@ -211,7 +217,7 @@ abstract class AbstractQuery extends BaseObject
    * @param string $sort
    * @param string $direction
    */
-  public function setSort(string $sort, string $direction) {
+  public function setSort(string $sort, string $direction): void {
     if ($this->allowSort($sort)) {
       $this->sort = $sort;
       $this->sortDirection = $direction == 'asc' ? 'asc' : 'desc';
@@ -224,9 +230,9 @@ abstract class AbstractQuery extends BaseObject
 
   /**
    * @param array $params
-   * @return array|mixed
+   * @return array
    */
-  protected function applyParameters($params = []) {
+  protected function applyParameters(array $params = []): array {
     if (!isset($params['sort']) && !empty($this->sort)) {
       $params['sort'] = $this->sort;
     }
@@ -250,12 +256,12 @@ abstract class AbstractQuery extends BaseObject
   /**
    * @return ElementQuery
    */
-  abstract protected function createQuery();
+  abstract protected function createQuery(): ElementQuery;
 
   /**
    * @param ElementQuery $query
    */
-  protected function setQuery(ElementQuery $query) {
+  protected function setQuery(ElementQuery $query): void {
     foreach ($this->_filters as $filter) {
       $filter->prepareQuery($this, $query);
     }
@@ -270,7 +276,7 @@ abstract class AbstractQuery extends BaseObject
   /**
    * @param Request $request
    */
-  protected function setRequest(Request $request) {
+  protected function setRequest(Request $request): void {
     foreach ($this->_filters as $filter) {
       $filter->setRequest($request);
     }
@@ -289,7 +295,7 @@ abstract class AbstractQuery extends BaseObject
    * @param array $config
    * @return static
    */
-  static public function create(array $config = []) {
+  static public function create(array $config = []): static {
     return new static($config);
   }
 }
