@@ -3,7 +3,6 @@
 namespace lenz\craft\essentials\assets;
 
 use craft\web\AssetBundle;
-use craft\web\assets\cp\CpAsset;
 
 /**
  * Class HandsontableAsset
@@ -13,7 +12,7 @@ class HandsontableAsset extends AssetBundle
   /**
    * @return void
    */
-  public function init() {
+  public function init(): void {
     $this->sourcePath = __DIR__ . '/handsontable';
     $this->js         = [ 'handsontable.full.min.js' ];
     $this->css        = [ 'handsontable.full.min.css' ];
