@@ -13,7 +13,7 @@ class EntryLoader extends AbstractLoader
   /**
    * @inheritDoc
    */
-  public function load(ElementInterface $element) {
+  public function load(ElementInterface $element): array|false {
     if (!($element instanceof Entry)) {
       return false;
     }
